requirements.txt
numpy==1.17.2
Pillow==6.1.0
scipy==1.3.1
tensorflow==1.14.0
